+++ src/__mock__/componentTest/quotes.test.js
